refactor(vst3): tidy processor state tests and fix compile issues

+++ examples/minimal-vst3/src/processor.rs
@@ -111,7 +111,7 @@ impl IComponentTrait for GainProcessor {
         bus.channelCount = 2;
         copy_wstring(label, &mut bus.name);
         bus.busType = BusTypes_::kMain as BusType;
-        bus.flags = BusInfo_::BusFlags_::kDefaultActive as u32;
+        bus.flags = BusInfo_::BusFlags_::kDefaultActive;
 
         kResultOk
     }
